# managed/src/main/resources/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "ALPHA - NOT FOR EXTERNAL USE",
    "version" : "v1",
    "title" : "Yugabyte Platform APIs",
    "termsOfService" : "TODO(chirag)",
    "contact" : {
      "name" : "https://docs.yugabyte.com"
    },
    "license" : {
      "name" : "Polyform Free Trial License 1.0.0",
      "url" : "https://github.com/yugabyte/yugabyte-db/blob/master/licenses/POLYFORM-FREE-TRIAL-LICENSE-1.0.0.txt"
    }
  },
  "host" : "localhost",
  "basePath" : "/",
...
  "schemes" : [ "http", "https" ],
  "consumes" : [ "application/json" ],
  "produces" : [ "application/json" ],
  "paths" : {
...
    "/api/v1/customers/{cUUID}/zones/{azUUID}/nodes/list" : {
      "get" : {
        "tags" : [ "Node Instances" ],
        "summary" : "List of node instances by zone",
        "description" : "",
        "operationId" : "listByZone",
        "parameters" : [ {
          "name" : "cUUID",
          "in" : "path",
          "required" : true,
          "type" : "string",
          "format" : "uuid"
        }, {
          "name" : "azUUID",
          "in" : "path",
          "required" : true,
          "type" : "string",
          "format" : "uuid"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/NodeInstance"
              }
            }
          }
        },
        "security" : [ {
          "apiKeyAuth" : [ ]
        } ]
      }
    },
    "/api/v1/customers/{cUUID}/providers/{pUUID}/nodes/list" : {
      "get" : {
        "tags" : [ "Node Instances" ],
        "summary" : "List of node instances by provider",
        "description" : "",
        "operationId" : "listByProvider",
        "parameters" : [ {
          "name" : "cUUID",
          "in" : "path",
          "required" : true,
          "type" : "string",
          "format" : "uuid"
        }, {
          "name" : "pUUID",
          "in" : "path",
          "required" : true,
          "type" : "string",
          "format" : "uuid"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/NodeInstance"
              }
            }
          }
        },
        "security" : [ {
          "apiKeyAuth" : [ ]
        } ]
      }
    },
    "/api/v1/customers/{cUUID}/providers/{pUUID}/instances/{instanceIP}" : {
      "delete" : {
        "tags" : [ "Node Instances" ],
        "summary" : "Delete node instance by ip",
        "description" : "",
        "operationId" : "deleteInstance",
        "parameters" : [ {
          "name" : "cUUID",
          "in" : "path",
          "required" : true,
          "type" : "string",
          "format" : "uuid"
        }, {
          "name" : "pUUID",
          "in" : "path",
          "required" : true,
          "type" : "string",
          "format" : "uuid"
        }, {
          "name" : "instanceIP",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        },
        "security" : [ {
          "apiKeyAuth" : [ ]
        } ]
      }
    },
    "/api/v1/customers/{cUUID}/universes/{universeUUID}/nodes/{nodeName}" : {
      "put" : {
        "tags" : [ "Node Instances" ],
        "summary" : "Update node actions",
        "description" : "",
        "operationId" : "nodeAction",
        "parameters" : [ {
          "name" : "cUUID",
          "in" : "path",
          "required" : true,
          "type" : "string",
          "format" : "uuid"
        }, {
          "name" : "universeUUID",
          "in" : "path",
          "required" : true,
          "type" : "string",
          "format" : "uuid"
        }, {
          "name" : "nodeName",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "Node action",
          "description" : "Node action data to be updated",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/NodeActionFormData"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/YWTask"
            }
          }
        },
        "security" : [ {
          "apiKeyAuth" : [ ]
        } ]
      }
    },
...
  },
  "securityDefinitions" : {
    "apiKeyAuth" : {
      "description" : "Api key passed as header",
      "type" : "apiKey",
      "name" : "X-AUTH-YW-API-TOKEN",
      "in" : "header"
    }
  },
  "definitions" : {
...
    "YWTask" : {
      "type" : "object",
      "required" : [ "taskUUID" ],
      "properties" : {
        "taskUUID" : {
          "type" : "string",
          "format" : "uuid"
        },
        "resourceUUID" : {
          "type" : "string",
          "format" : "uuid",
          "description" : "UUID of the resource being modified  by the task",
          "readOnly" : true
        }
      }
    },
...
    "NodeInstance" : {
      "type" : "object",
      "required" : [ "details" ],
      "properties" : {
        "nodeUuid" : {
          "type" : "string",
          "format" : "uuid",
          "description" : "Node instance UUID",
          "readOnly" : true
        },
        "instanceTypeCode" : {
          "type" : "string",
          "example" : "c5large",
          "description" : "Node instance type code"
        },
        "nodeName" : {
          "type" : "string",
          "example" : "India node",
          "description" : "Node instance node name"
        },
        "instanceName" : {
          "type" : "string",
          "example" : "Mumbai instance",
          "description" : "Node instance instance name"
        },
        "zoneUuid" : {
          "type" : "string",
          "format" : "uuid",
          "description" : "Node instance zone UUID"
        },
        "inUse" : {
          "type" : "boolean",
          "description" : "Node instance is in used or not"
        },
        "details" : {
          "$ref" : "#/definitions/NodeInstanceData"
        },
        "detailsJson" : {
          "type" : "string",
          "example" : "{\"ip\":\"1.1.1.1\",\"sshUser\":\"centos\"}",
          "description" : "Node details"
        }
      },
      "description" : "Node instance attched to provider and zones"
    },
    "NodeInstanceData" : {
      "type" : "object",
      "required" : [ "instanceName", "instanceType", "ip", "region", "sshUser", "zone" ],
      "properties" : {
        "ip" : {
          "type" : "string",
          "example" : "1.1.1.1",
          "description" : "IP address of node instance"
        },
        "sshUser" : {
          "type" : "string",
          "example" : "centos",
          "description" : "SSH user of node instance"
        },
        "region" : {
          "type" : "string",
          "example" : "south-east",
          "description" : "Region of node instance"
        },
        "zone" : {
          "type" : "string",
          "example" : "south-east",
          "description" : "Zone of node instance"
        },
        "instanceType" : {
          "type" : "string",
          "example" : "c5large",
          "description" : "Instance type of node instance"
        },
        "instanceName" : {
          "type" : "string",
          "example" : "Mumbai instance",
          "description" : "Instance name of node instance"
        },
        "nodeName" : {
          "type" : "string",
          "example" : "India node",
          "description" : "Node name of node instance"
        }
      },
      "description" : "Detail of node instance"
    },
...
    "NodeActionFormData" : {
      "type" : "object",
      "required" : [ "nodeAction" ],
      "properties" : {
        "nodeAction" : {
          "type" : "string",
          "enum" : [ "ADD", "REMOVE", "START", "STOP", "DELETE", "QUERY", "RELEASE", "START_MASTER" ]
        }
      }
    },
...
  },
...
}